pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "twincart-product-matching"
version = "1.0.0"
description = "TwinCart: Multimodal product matching for E-commerce."
readme = "README.md"
requires-python = ">=3.11,<3.13"
license = { file = "LICENSE" }
authors = [{ name = "Andryushenko Solomon" }]
dependencies = [
  # core
  "numpy>=1.26",
  "pandas>=2.2",
  "pyyaml>=6.0.2",
  "tqdm>=4.66",
  "rich>=13.7",

  # hydra
  "hydra-core>=1.3.2",
  "omegaconf>=2.3.0",

  # matching
  "faiss-cpu>=1.8.0",

  # dvc
  "dvc>=3.50",
  "dvc-gdrive>=3.0.0",
]

[project.optional-dependencies]
dev = [
  "pre-commit>=3.7",
  "pytest>=8.3",
  "pytest-cov>=5.0",
  "black>=24.4",
  "isort>=5.13",
  "flake8>=7.0",
  "ruff>=0.5.0",
]

train = [
  # torch stack
  "torch>=2.3",
  "torchvision>=0.18",
  "pytorch-lightning>=2.3",
  "timm>=1.0.7",

  # nlp
  "transformers>=4.43",
  "tokenizers>=0.19",

  # metrics & optimizer
  "pytorch-metric-learning>=2.5",
  "madgrad>=1.3",

  # augmentation
  "albumentations>=1.4",
  "opencv-python-headless>=4.10",

  # utilities
  "scikit-learn>=1.5",
  "pyarrow>=16.0",
  "matplotlib",
]

logging = [
  "mlflow>=2.14",
]

export = [
  "onnx>=1.16",
  "onnxruntime>=1.18",
  "onnxscript>=0.5.7",
]

serve = [
  "grpcio>=1.64",
  "tritonclient[grpc]>=2.50.0,<2.58.0",
  "gunicorn>=21.2",
  "transformers>=4.43",
  "tokenizers>=0.19",
  "mlflow>=2.14",
  "onnxruntime>=1.18",
  "pillow>=10.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.black]
line-length = 160
target-version = ["py311"]

[tool.isort]
profile = "black"
line_length = 160
src_paths = ["src", "tests"]

[tool.flake8]
max-line-length = 160
extend-ignore = [
  "E203",
  "W503",
]
exclude = [
  ".git",
  ".venv",
  "outputs",
  "multirun",
  "mlruns",
  "data",
  "models",
]

[tool.ruff]
line-length = 160
target-version = "py311"
extend-exclude = ["outputs", "multirun", "mlruns", "data", "models"]
lint.select = ["E", "F", "I", "B", "UP"]
lint.ignore = ["E501", "I001", "E301", "E302", "E303", "E305", "E306"]

[tool.ruff.lint.isort]
force-sort-within-sections = false

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q --disable-warnings --maxfail=1"

[dependency-groups]
serve = [
    "tritonclient[grpc]>=2.57.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[project.scripts]
twincart = "twincart.commands:main"
twincart-prepare-folds = "twincart.commands:prepare_folds"
twincart-train = "twincart.commands:train"
twincart-build-mlflow = "twincart.serving.build_mlflow:main"
